pyproject.toml
[project]
name = "rat"
version = "0.1.0"
description = "A powerful AI reasoning chain CLI tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.44.0",
    "colorama>=0.4.6",
    "openai>=1.59.9",
    "prompt-toolkit>=3.0.50",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "flask>=3.0.0",
]

[project.scripts]
rat = "rat.rat:main"
rat-web = "web.app:app.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rat", "web"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=rat --cov-report=term-missing"
